backend/resume-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic[email]==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
redis[asyncio]==5.2.0
aio-pika==9.4.3
httpx==0.27.2
openai==1.58.1
motor==3.6.0
python-multipart==0.0.12
PyPDF2==3.0.1
python-docx==1.1.2
